requirements.txt
pyspark==3.5.6
delta-spark==3.3.0
dbt-core
dbt-spark[PyHive]
python-dotenv
pandas
kafka-python
psycopg2-binary
pyarrow
fastparquet
grpcio
grpcio-tools
git+https://github.com/Automated-Datavault-Schema-Evolution/MyLogger.git@v0.2.1#egg=mylogger
